actors as independent widget
