+++ lib/presentation/screens/movies/movie_creen.dart
@@ -111,7 +111,7 @@ class _MovieDetail extends StatelessWidget {
             ],
           ),
         ),
-        _ActorsByMovie(
+        ActorsByMovie(
           movieId: movie.id,
         ),
         Padding(
@@ -141,70 +141,6 @@ class _MovieDetail extends StatelessWidget {
     );
   }
 }
-
-class _ActorsByMovie extends ConsumerWidget {
-  final int movieId;
-  const _ActorsByMovie({required this.movieId});
-
-  @override
-  Widget build(BuildContext context, ref) {
-    final actorsByMovie = ref.watch(actorsByMovieProvider);
-
-    if (actorsByMovie[movieId.toString()] == null) {
-      return const CircularProgressIndicator(
-        strokeWidth: 2,
-      );
-    }
-
-    final actors = actorsByMovie[movieId.toString()]!;
-
-    return SizedBox(
-      height: 285,
-      child: ListView.builder(
-        scrollDirection: Axis.horizontal,
-        itemCount: actors.length,
-        itemBuilder: (context, index) {
-          final actor = actors[index];
-
-          return Container(
-            padding: const EdgeInsets.all(8.0),
-            width: 135,
-            child:
-                Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
-              FadeInRight(
-                child: ClipRRect(
-                  borderRadius: BorderRadius.circular(20),
-                  child: Image.network(
-                    actor.profilePath,
-                    height: 180,
-                    width: 135,
-                    fit: BoxFit.cover,
-                  ),
-                ),
-              ),
-              const SizedBox(
-                height: 5,
-              ),
-              Text(
-                actor.name,
-                maxLines: 2,
-              ),
-              Text(
-                actor.character ?? "",
-                maxLines: 2,
-                style: const TextStyle(
-                    fontWeight: FontWeight.bold,
-                    overflow: TextOverflow.ellipsis),
-              ),
-            ]),
-          );
-        },
-      ),
-    );
-  }
-}
-
-
 class _CustomSliverAppBar extends ConsumerWidget {
   final Movie movie;
   const _CustomSliverAppBar({required this.movie});
